requirements.txt
tensorflow
keras
pandas
numpy
scikit-learn
matplotlib
joblib
pyyaml
